Adds back smartBorders
It does look kind of nice to not have borders when you only have one
screen on the workspace.

And this wasn't the cause of overlapping xmobar. Not sure what the cause
is but since I added trayer everything seems to work fine because it's
the same height as xmobar.