Replace diodon with greenclip (and rofi)
Got tired of g_object_set_data: assertion 'G_IS_OBJECT (object)' failed

Parcellite works but it's hard to read since it has no scaling on
high dpi screens.

greenclip is written in haskell and is available for download as a
static binary. It integrates well with rofi which is cool.