Migrate to new autocomplete plugin for vim
This one is powered by deno and so we had to install that locally.

Easiest is with `nix-env -f "<nixpkgs>" -iA deno`. I need to add this to
the home-manager config as well.

Reason for the change is because Shougo recommends migrating from
deoplete and because I got sick of vim-hug-neovim-rpc and it's inability
to set the python used - which required installing pynivm in each
virtualenv and/or nix shell.