xmonad: actually fixes the refocus issue
This was particularly annoying with gnome pinentry as it would
pop up during a ssh command (because I use yubikey for my
ssh key ... via gpg) and then XMonad would focus the wrong window when
pinentry closes. It happens with other popups ocasionally but this one
was the most annoying.