Disable hybrid layout and switch back to normal
As I said in the previous commit - I don't think this is for me
but want to keep it around for future reference.

I removed setting Xmodmap as I don't need that anymore seeing as
xkb is taking care of switching caps lock to control.
Xmodmap was only needed for xcape to function correctly.